requirements.txt
pandas
xlrd
flask
gunicorn
# uwsgi
Flask-SQLAlchemy
Flask-APScheduler
fastapi
uvicorn
slowapi
sqlalchemy_utils